refactor(TREDS-47): rename LoadingIndicators to LoadingSpinner
Aligns folder, file, and export names with the single-component convention
used by every other component in the library.
