--- src/components/LoadingIndicators/index.ts
@@ -1,1 +1,0 @@
-export * from './LoadingIndicators';
